refactor: split draft.ts into state/picks/queries modules
