+++ src/actions/draft/queries.ts
@@ -1,0 +1,4 @@
+"use server";
+
+// Query functions for draft state and draft pool will be added here.
+// These read-only functions provide draft data to the UI layer.
